+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Not released
 
 - Add `merge` and `mergeBlock` to `Authorizer` (#35) (Clément Delafargue)
+- Add an express middleware for proctecting endpoints (#35) (Clément Delafargue)
 
 # `0.4.0-beta2`